Add analytics tags for error pages and don't fire the pageLoad beacon on them
