+++ pages/404.tsx
@@ -60,6 +60,7 @@ export const getStaticProps: GetStaticProps = async ({ locale }) => {
       service: 'ESDC-EDSC_MSCA-MSDC-SCH',
       creator: 'Employment and Social Development Canada',
       accessRights: '1',
+      statusCode: '404',
     },
     data_fr: {
       title: `Nous ne pouvons trouver cette page Web - 404 - Mon dossier Service Canada`,
@@ -69,6 +70,7 @@ export const getStaticProps: GetStaticProps = async ({ locale }) => {
       service: 'ESDC-EDSC_MSCA-MSDC-SCH',
       creator: 'Emploi et Développement social Canada',
       accessRights: '1',
+      statusCode: '404',
     },
   }
   return {
