+++ pages/_app.js
@@ -8,9 +8,6 @@ import ErrorBoundary from '../components/ErrorBoundary'
 import { useEffect } from 'react'
 config.autoAddCss = false
 
-// To help prevent double firing of adobe analytics pageLoad event
-let appPreviousLocationPathname = ''
-
 export default function MyApp({ Component, pageProps, router }) {
   /* istanbul ignore next */
   if (Component.getLayout) {
@@ -19,27 +16,6 @@ export default function MyApp({ Component, pageProps, router }) {
 
   Modal.setAppElement('#__next')
   const display = { hideBanner: pageProps.hideBanner }
-
-  /** Web Analytics - taken from Google Analytics example
-   *  @see https://github.com/vercel/next.js/blob/canary/examples/with-google-analytics
-   * */
-  useEffect(() => {
-    const handleRouteChange = () => {
-      // only push event if pathname is different
-      if (window.location.pathname !== appPreviousLocationPathname) {
-        window.adobeDataLayer?.push?.({ event: 'pageLoad' })
-        appPreviousLocationPathname = window.location.pathname
-      }
-    }
-
-    handleRouteChange()
-    router.events.on('routeChangeComplete', handleRouteChange)
-    router.events.on('hashChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-      router.events.off('hashChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
 
   /* istanbul ignore next */
   return (
